requirements-dev.txt
flask
flask-httpauth
mysql-connector-python
gunicorn
pandas
openpyxl
requests
ttkbootstrap
pillow
matplotlib